Rename some variables to be more correct/understandable.
Summary:
In particular, I audited the source code to be more careful about how
I used the term `job` to refer to a Jenkins job, and make sure I was
using `build` when referring to a build (a particular run of a job).

I also changed references to a collection of jenkins builds to be a
"task" rather than a "deploy," since other clients may use this for
analyzing other tasks besides deploying.

I also renamed `timerange` to be `intervals`.  Having it be a singular
for a plural concept was very confusing.  I also made an Interval
named-tuple instead of having to do `[0]` and `[1]` and `[2]`
everywhere.

Finally, I fixed up the imports to work with both python2 and python3.

Test Plan:
I ran these with success:
   ./visualize_jenkins_perf_data.py deploy/build-webapp:14167
   ./visualize_jenkins_perf_data.py /tmp/jenkins-build-perf-analysis/deploy--build-webapp:14167.data
   python3 ./visualize_jenkins_perf_data.py /tmp/jenkins-build-perf-analysis/deploy--build-webapp:14167.data

Reviewers: benkraft

Reviewed By: benkraft

Differential Revision: https://phabricator.khanacademy.org/D66512