Add, and use, a configuration library.
Summary:
I kinda went overboard here, and support both a configuration json
file (with comments) and commandline-flag overrides.

The main thing I wanted to configure was the color use.  This ended up
resulting in quite a bit of code reorganization, but I like the end
result better.

Test Plan:
I ran the following, and they all showed the same graph (which also
looks the same as before this commit):
   ./visualize_jenkins_perf_data.py deploy/build-webapp:14167 -c jenkins_perf_visualizer/config.json
   python3 ./visualize_jenkins_perf_data.py deploy/build-webapp:14167 -c jenkins_perf_visualizer/config.json
   ./visualize_jenkins_perf_data.py /tmp/jenkins-build-perf-analysis/deploy--build-webapp:14167.data -c jenkins_perf_visualizer/config.json
   python3 ./visualize_jenkins_perf_data.py /tmp/jenkins-build-perf-analysis/deploy--build-webapp:14167.data -c jenkins_perf_visualizer/config.json

I then ran:
   ./download_jenkins_perf_data.py deploy/build-webapp -c jenkins_perf_visualizer/config.json
and saw from the jumbled stdout that it was correctly multithreading.

I hit control-C (actually control-\) halfway through, and did
   ls -l /tmp/jenkins-build-perf-analysis/
to ensure the data files were being downloaded with non-zero size.
I then did
   ./download_jenkins_perf_data.py deploy/build-webapp -c jenkins_perf_visualizer/config.json --download-threads=1
which both tests the commandline-flag override, and also the
single-threaded codepath (which is new).  It worked as expected.

At the end of the run I did
   ls /tmp/jenkins-build-perf-analysis/
and saw a bunch of directories like:
   2020-10.74629/
That directory had two files (symlinks) in it, which is expected since
a single deploy usually builds at least twice.  I did
   cat /tmp/jenkins-build-perf-analysis/2020-10.74629/deploy--build-webapp:14162.data
and it gave content, showing the symlink is set up correctly.

Reviewers: benkraft

Reviewed By: benkraft

Differential Revision: https://phabricator.khanacademy.org/D66513