Add the ability to override the title from the commandline.
Normally we figure it out automatically from the jenkins builds being
graphed.  But sometimes you have a better idea of what to do on a
per-graph basis.  This adds a commandline flag to let you do that.

I also changed the UI a bit to separate out the title from the
subtitle (when the graph started basically).  The new `--title` flag
only overrides the former; the latter is (currently) fixed.

Auditors: benkraft